Vehicle/vehicle_serializers.py
from rest_framework import serializers
from Vehicle.models import Vehicle

class VehicleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Vehicle
        fields = ['id', 'brand', 'model', 'year']