requirements.txt
asgiref==3.8.1
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
Django==4.2.11
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
inflection==0.5.1
outcome==1.3.0.post0
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PySocks==1.7.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
selenium==4.20.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wsproto==1.2.0